Add "Export Data" button, which shows the geoJSON in a new window.
Putting modified data in a new browser window is about the only thing
a "save" button can do in a purely client-side framework; it's then up
to the user to put it where it can do some good.

Problems:
  *) button needs better placement
  *) button should be suppressed if editing not configured
  *) generated output is not pretty-printed, and fugly as hell.

Still, going for proof of concept here, and this is it.