In edit mode, currently selected point is draggable.
Furthermore, dragging it alters the stored geoJson, and is correctly
reflected on export, all managed by the edit_state component (which is
listening for "selectFeature" and "selectedFeatureMoved" events).

Works when tested manually, and verifies the general theory of edits for
other stuff (and how I expect edit state to be bound to the UI, at least in
general terms).  Automated tests tbd.