Double-clicking on the map offers to add a point.
This is actually done in two stages:  the double-click gives you a popup
in which you enter the name of the feature you're creating (well, the
map's preview feature, whatever that is); submitting the form actually
creates the feature, and gives you the full feature editor.  To minimize
mousing-around, the popup form automatically gets keyboard focus when
created.

The intent of this dance is to make a stray double-click easy to undo.

Internally, what happens is that on submission, the popup form handler
sends a 'newFeature' event, which anything maintaining its own copy of
the data ought to handle.  The edit-state does that by adding the feature
to its data.  The map adds a new marker.  (The map is also, currently, what
sends the event, but you can imagine cases where it wouldn't be.)