Make sure we can still find a feature's marker after a position change.
For this purpose, map.js uses a map indexed by the [lng, lat] coordinate
arrays from the feature geometries.  And the hash lookups for those, in
turn, do depend on the coordinate values.  So, at least as a stopgap, we
now re-record the feature-to-marker mapping in this map after drag-end.