Mitigate Heisenbug in map specs.
Leaflet sets timeouts internally.  These can fire after the test that
created the map is over, and the map is no longer present in the DOM.

That doesn't always go well.

This patch adds an afterEach to the map specs, which effectively
defangs a timeout for which it can go rather badly -- throwing an
exception in the timeout, which causes an early abort from the whole
test run when the exception goes uncaught.  See comment in the patch
for details.