Implement rudimentary "undo" capability.
Specifically, this gives the edit-info pane a "revert" button which
restores the feature's position, properties, and delete status to whatever
they were on first load.  The edit-state component is in charge of managing
undo status, which it does by stuffing stuff into the "undoInfo" of
features (a member which no one else is supposed to look at).

Other components are made aware of the undo status of the selected feature
(if they care) by listening for the 'selectedFeatureUndoStatus' event,
which carries a boolean value (true if "revert" will do something, false
if the values are unchanged).

Other components can request an undo by sending the "requestUndo" event,
which causes the edit state to fish out the original values, and send
"selectedFeaturePropsChanged", "selectedFeatureMoved", and undelete
events to restore the original values.  (As a consequence, the infoedits
component is now listening for prop-change events, so that it shows the
reverted values when a revert occurs!)

This protocol is designed so that if we want to go to an actual per-event
undo stack, nothing needs to change outside of the edit state (though we
might want to relabel the revert button!).  For map-wide undo state, the
protocol would have to change somewhat, but then again, so would the UI;
it would no longer be appropriate to have the undo-triggering button on
a control tied to a particular feature.