In edit mode, allow users to create a point by entering the address.
In edit mode only, this creates a second submit button on the geosearch
form (if that is enabled!), which initiates the process of creating a new
feature at the location, if the geosearch result shows a location.

This is an alternative to double-clicking on the map, which may be
quicker in some cases -- users will have to type the address anyway,
and this way, they can be sure of the exact spot.  (Or, well, at least
as sure as they are of the geolocation server.)